context/activation_analysis/requirements.txt
torch>=2.1.0
transformers>=4.40.0
pyyaml>=6.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
tqdm>=4.65.0
numpy>=1.24.0
scikit-learn>=1.3.0
